% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac.resid.spca.R
\name{ac.resid.spca}
\alias{ac.resid.spca}
\title{Residualized adjusted for confounding supervised principal component analysis}
\usage{
ac.resid.spca(
  X,
  Y,
  a,
  L,
  A,
  H,
  lambda,
  npc,
  fast.mat = F,
  parallel = F,
  clust = NULL
)
}
\arguments{
\item{X}{nxp data matrix}

\item{Y}{nx1 response matrix}

\item{a}{nxq confounding variable matrix}

\item{L}{kernel matrix of response}

\item{A}{kernel matrix of confounders}

\item{H}{centering matrix}

\item{lambda}{penalty term for confounding variable adjustment}

\item{npc}{number of principl components}

\item{fast.mat}{flag for whether to use Rfast coding}

\item{parallel}{parallelization of residualization}

\item{clust}{cluster of cores for parallel procedure}
}
\description{
Residualized adjusted for confounding supervised principal component analysis
}
\examples{
ac.resid.spca(X, Y, a, L, A, H, lambda, npc, fast.mat=F, parallel=F, clust=NULL)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
