% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsimpls.R
\name{acsimpls}
\alias{acsimpls}
\title{SIMPLS algorithm included adjustment for confounding variables}
\usage{
acsimpls(
  X,
  Y,
  A,
  lambda,
  npc,
  resp.kernel = c("linear", "delta"),
  conf.kernel = c("linear", "gaussian"),
  bandwidth = NULL
)
}
\arguments{
\item{X}{nxp predictor matrix}

\item{Y}{nx1 response vector}

\item{A}{nxq confounding variable matrix}

\item{resp.kernel}{kernel of response vector which can be linear or delta}

\item{conf.kernel}{kernel of confounding matrix which can be linear or gaussian}

\item{bandwidth}{gaussian kernel argument}
}
\description{
SIMPLS algorithm included adjustment for confounding variables
}
\examples{
acsimpls(X,Y,A,lambda,npc,resp.kernel="linear", conf.kernel="linear", bandwidth=NULL)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
