% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acSPCA.R
\name{acSPCA}
\alias{acSPCA}
\title{Supervised Adjusted for Confounding Principal Component Analysis}
\usage{
acSPCA(
  X,
  Y,
  A,
  lambda,
  npc,
  resp.kernel = c("linear", "delta"),
  conf.kernel = c("linear", "gaussian"),
  bandwidth = NULL
)
}
\arguments{
\item{X}{nxp predictor matrix}

\item{Y}{nx1 response vector}

\item{A}{nxq confounding variable matrix}

\item{lambda}{penalty to adjust for confounding}

\item{npc}{number of desired principal component}

\item{resp.kernel}{kernel of response vector which can be linear or delta}

\item{conf.kernel}{kernel of confounding matrix which can be linear or gaussian}

\item{bandwidth}{gaussian kernel argument}
}
\description{
Supervised Adjusted for Confounding Principal Component Analysis
}
\examples{
acSSPCA(X, Y, A, X4A=NULL, X4Y=NULL, c1=NULL, c2, v_ini, resp.kernel=c("linear","delta"), conf.kernel=c("linear", "gaussian"), bandwidth=NULL, maxiter=50, delta=10^-8, filter=T, minmaxsep=1e-3)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Supervised}
