% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acSSPCA.R
\name{acSSPCA}
\alias{acSSPCA}
\title{Sparse Supervised Adjusted for Confounding Principal Component Analysis}
\usage{
acSSPCA(
  X,
  Y,
  A,
  X4A = NULL,
  X4Y = NULL,
  c1 = NULL,
  c2,
  v_ini,
  resp.kernel = c("linear", "delta"),
  conf.kernel = c("linear", "gaussian"),
  bandwidth = NULL,
  maxiter = 50,
  delta = 10^-8,
  filter = T,
  minmaxsep = 0.001
)
}
\arguments{
\item{X}{nxp predictor matrix}

\item{Y}{nx1 response vector}

\item{A}{nxq confounding variable matrix}

\item{X4A}{alternative confounding matrix starting point}

\item{c1}{norm of quadratic predictor and confounding matrix product}

\item{c2}{sum of absolute value loadings for desired sparse vector}

\item{v_ini}{non-sparse initial vector}

\item{resp.kernel}{kernel of response vector which can be linear or delta}

\item{conf.kernel}{kernel of confounding matrix which can be linear or gaussian}

\item{bandwidth}{gaussian kernel argument}

\item{maxiter}{maximum number of iterations for the algorithm}

\item{delta}{offset parameter}

\item{filter}{parameter to determine filtering}

\item{minmaxsep}{determines the acceptable algorithm termination acceptable different}

\item{x4Y}{alternative response variable starting point}
}
\description{
Sparse Supervised Adjusted for Confounding Principal Component Analysis
}
\examples{
acSSPCA(X, Y, A, X4A=NULL, X4Y=NULL, c1=NULL, c2, v_ini, resp.kernel=c("linear","delta"), conf.kernel=c("linear", "gaussian"), bandwidth=NULL, maxiter=50, delta=10^-8, filter=T, minmaxsep=1e-3)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
