% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acSSPCAmulti.R
\name{acSSPCAmulti}
\alias{acSSPCAmulti}
\title{Sparse Supervised Adjusted for Confounding Principal Component Analysis allowing multiple sparse principal components}
\usage{
acSSPCAmulti(
  X,
  Y,
  A,
  lambda,
  npc,
  resp.kernel = c("linear", "delta"),
  conf.kernel = c("linear", "gaussian"),
  bandwidth = NULL,
  c1 = NULL,
  c2 = NULL,
  maxiter = 50,
  delta = 10^-8,
  filter = T,
  minmaxsep = 0.001
)
}
\arguments{
\item{X}{nxp predictor matrix}

\item{Y}{nx1 response vector}

\item{A}{nxq confounding variable matrix}

\item{resp.kernel}{kernel of response vector which can be linear or delta}

\item{conf.kernel}{kernel of confounding matrix which can be linear or gaussian}

\item{bandwidth}{gaussian kernel argument}

\item{c1}{norm of quadratic predictor and confounding matrix product}

\item{c2}{sum of absolute value loadings for desired sparse vector}

\item{maxiter}{maximum number of iterations for the algorithm}

\item{delta}{offset parameter}

\item{filter}{parameter to determine filtering}

\item{minmaxsep}{determines the acceptable algorithm termination acceptable different}
}
\description{
Sparse Supervised Adjusted for Confounding Principal Component Analysis allowing multiple sparse principal components
}
\examples{
acSSPCAmulti(X,Y,A,lambda,npc,resp.kernel=c("linear","delta"), conf.kernel=c("linear", "gaussian"), bandwidth=NULL, c1=NULL, c2=NULL, maxiter=50, delta=10^-8, filter=T, minmaxsep=1e-3)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
