% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calkernel.R
\name{calkernel}
\alias{calkernel}
\title{Calculates the kernel for the confounding variables}
\usage{
calkernel(Y, kernel, bandwidth, scaleY = F)
}
\arguments{
\item{Y}{Confounding variable matrix}

\item{kernel}{Type of kernel.  Either linear or gaussian}

\item{bandwidth}{scaling parameter for gaussian kernel}

\item{scaleY}{scales the confounding variables}
}
\description{
Calculates the kernel for the confounding variables
}
\examples{
respkernel(Y, n, kernel)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
