% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.acSSPCA.R
\name{cv.acSSPCA}
\alias{cv.acSSPCA}
\title{CV function for determining sparsity penalty c2}
\usage{
cv.acSSPCA(
  X,
  Y,
  A,
  lambda,
  npc,
  resp.kernel = c("linear", "delta"),
  conf.kernel = c("linear", "gaussian"),
  bandwidth = NULL,
  c1 = NULL,
  c2 = NULL,
  n.folds = 5,
  maxiter = 50,
  delta = 10^-8,
  filter = T,
  minmaxsep = 0.001,
  n.cores = 4,
  parallel = F
)
}
\arguments{
\item{X}{nxp predictor matrix}

\item{Y}{nx1 response vector}

\item{A}{nxq confounding variable matrix}

\item{lambda}{penalty for adjusting for confounding}

\item{npc}{number of principal components}

\item{resp.kernel}{kernel of response vector which can be linear or delta}

\item{conf.kernel}{kernel of confounding matrix which can be linear or gaussian}

\item{bandwidth}{gaussian kernel argument}

\item{c1}{norm of quadratic predictor and confounding matrix product}

\item{c2}{vector of values given as sum of absolute value loadings for desired sparse vector}

\item{maxiter}{maximum number of iterations for the algorithm}

\item{delta}{offset parameter}

\item{filter}{parameter to determine filtering}

\item{minmaxsep}{determines the acceptable algorithm termination acceptable different}

\item{n.cores}{number of cores for parallel option}

\item{parallel}{flag for whether or not parallelization should be used}
}
\description{
CV function for determining sparsity penalty c2
}
\examples{
cv.acSSPCA <- function(X,Y,A,lambda,npc,resp.kernel=c("linear","delta"), conf.kernel=c("linear","gaussian"), bandwidth=NULL, c1=NULL, c2=NULL, n.folds=5, maxiter=50, delta=10^-8, filter=T, minmaxsep=1e-3, n.cores=4, parallel=F)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
