% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.partition.acSSPCA.lambda.R
\name{cv.partition.acSSPCA.lambda}
\alias{cv.partition.acSSPCA.lambda}
\title{CV function for determining sparsity penalty c2}
\usage{
cv.partition.acSSPCA.lambda(
  arg.sparse,
  df.partition,
  npc,
  n.folds,
  resp.kernel = c("linear", "delta"),
  conf.kernel = c("linear", "gaussian"),
  bandwidth = NULL,
  c1 = NULL,
  maxiter = 50,
  delta = 10^-8,
  filter = T,
  minmaxsep = 0.001,
  x.names = NULL,
  y.names = NULL,
  a.names = NULL
)
}
\arguments{
\item{arg.sparse}{vector for folds and sparse parameter}

\item{df.partition}{list of folds}

\item{npc}{number of principal components}

\item{n.folds}{number of folds for cross-validation}

\item{resp.kernel}{kernel of response vector which can be linear or delta}

\item{conf.kernel}{kernel of confounding matrix which can be linear or gaussian}

\item{bandwidth}{gaussian kernel argument}

\item{c1}{norm of quadratic predictor and confounding matrix product}

\item{maxiter}{maximum number of iterations for the algorithm}

\item{delta}{offset parameter}

\item{filter}{parameter to determine filtering}

\item{minmaxsep}{determines the acceptable algorithm termination acceptable different}

\item{x.names}{column names for predictors}

\item{y.names}{name of response variable}

\item{a.names}{name of confounding variables}
}
\description{
CV function for determining sparsity penalty c2
}
\examples{
cv.acSSPCA <- function(X,Y,A,lambda,npc,resp.kernel=c("linear","delta"), conf.kernel=c("linear","gaussian"), bandwidth=NULL, c1=NULL, c2=NULL, n.folds=5, maxiter=50, delta=10^-8, filter=T, minmaxsep=1e-3, n.cores=4, parallel=F)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
