% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflate.rfast.R
\name{deflate.rfast}
\alias{deflate.rfast}
\title{Deflation of the X matrix in the SIMPLS algorithm}
\usage{
deflate.rfast(X, V, p)
}
\arguments{
\item{X}{nxp predictor matrix}

\item{V}{matrix containing principal component vectors}

\item{p}{number of predictors in data matrix X}
}
\description{
Deflation of the X matrix in the SIMPLS algorithm
}
\examples{
deflate.rfast(X,V,p)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
