% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.psi.R
\name{get.psi}
\alias{get.psi}
\title{Calculates the Psi matix Psi=Delta^t HX.}
\usage{
get.psi(X, L, n, p)
}
\arguments{
\item{X}{nxp data matrix.  Typically should be centered}

\item{L}{nxn response kernel matrix}

\item{n}{number of responses in the data}

\item{p}{number of predictors in the data}
}
\description{
Calculates the Psi matix Psi=Delta^t HX.
}
\examples{
get.psi(X,L,n,p)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
