% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.resid.apply.R
\name{get.resid.apply}
\alias{get.resid.apply}
\title{Uses apply function to perform residual calculation}
\usage{
get.resid.apply(X, Y, a)
}
\arguments{
\item{X}{nxp predictor matrix}

\item{Y}{nx1 response vector}

\item{a}{nxq confounding matrix}
}
\description{
Uses apply function to perform residual calculation
}
\examples{
get.resid.apply(X,Y,a)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
