% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.resid.parallel.R
\name{get.resid.parallel}
\alias{get.resid.parallel}
\title{Uses parApply function to perform residual calculation}
\usage{
get.resid.parallel(X, Y, a, clust)
}
\arguments{
\item{X}{nxp predictor matrix}

\item{Y}{nx1 response vector}

\item{a}{nxq confounding matrix}

\item{clust}{cluster of cores for parallelization}
}
\description{
Uses parApply function to perform residual calculation
}
\examples{
get.resid.parallel(X, Y, a, clust)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
