% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.resid.single.R
\name{get.resid.single}
\alias{get.resid.single}
\title{For loop residualization}
\usage{
get.resid.single(X, Y, a)
}
\arguments{
\item{X}{nxp predictor matrix}

\item{Y}{nx1 response vector}

\item{a}{nxq confounding matrix}
}
\description{
For loop residualization
}
\examples{
get.resid.single(X,Y,a)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
