% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.residual.R
\name{get.residual}
\alias{get.residual}
\title{Residualization function}
\usage{
get.residual(X, Y, a, parallel = F, clust = NULL)
}
\arguments{
\item{X}{nxp predictor matrix}

\item{Y}{nx1 response vector}

\item{a}{nxq confounding matrix}

\item{parallel}{flag for whether to use apply or parApply}

\item{clust}{cluster of cores for parallization}
}
\description{
Residualization function
}
\examples{
get.residual(X, Y, a, parallel=F, clust=NULL)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
