% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnorm2.R
\name{getnorm2}
\alias{getnorm2}
\title{L2 norm calculation}
\usage{
getnorm2(vector)
}
\arguments{
\item{vector}{vector to calculate l2 norm}
}
\description{
L2 norm calculation
}
\examples{
getnorm2(vector)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
