% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj}
\alias{proj}
\title{Internal function to perform L1 penalization on vector}
\usage{
proj(M, S, V, utx, c1, c2, v_ini, tmin, tmax, iternum, minmaxsep)
}
\arguments{
\item{M}{Confounding matrix product with predictors}

\item{S}{Singular values of M matrix}

\item{V}{right vectors of M matrix}

\item{utx}{crossproduct of u vector and Psi matrix}

\item{c2}{sum of absolute value loadings for desired sparse vector}

\item{v_ini}{initial vector to undergo L1 penalization and projection}

\item{tmin}{iterative lower bound}

\item{tmax}{iterative upper bound}

\item{iternum}{number of iterations}

\item{minmaxsep}{determines the acceptable algorithm termination acceptable different}
}
\description{
Internal function to perform L1 penalization on vector
}
\examples{
proj( M, S, V, utx, c1, c2, v_ini, tmin, tmax, iternum, minmaxsep)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
