% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_hp.R
\name{proj_hp}
\alias{proj_hp}
\title{Projection onto hyperplane}
\usage{
proj_hp(v, w, t)
}
\arguments{
\item{v}{vector to under go projection}

\item{w}{weight vector}

\item{t}{iterative value}
}
\description{
Projection onto hyperplane
}
\examples{
proj_hp( v, w, t)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
