% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_l1.R
\name{proj_l1}
\alias{proj_l1}
\title{Internal function to perform l1 penalization}
\usage{
proj_l1(v, c)
}
\arguments{
\item{v}{vector to under go l1 penalty}

\item{c}{l1 penalty term}
}
\description{
Internal function to perform l1 penalization
}
\examples{
proj_l1(v, c)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
