% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_l2.R
\name{proj_l2}
\alias{proj_l2}
\title{Internal function to perform l2n penalization on vector}
\usage{
proj_l2(v, c)
}
\arguments{
\item{v}{vector to under go l2 norm normalization}

\item{c}{constant to multiply l2n norm 1 vector by}
}
\description{
Internal function to perform l2n penalization on vector
}
\examples{
proj_l2function(v, c)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
