% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_l2g_getl.R
\name{proj_l2g_getl}
\alias{proj_l2g_getl}
\title{Projection onto hyperplane}
\usage{
proj_l2g_getl(S, P, c)
}
\arguments{
\item{S}{Singular values}

\item{P}{Projection matrix for v}

\item{c}{norm of crossprod between predictor matrix and confounding matrix}
}
\description{
Projection onto hyperplane
}
\examples{
proj_l2g_getl( S, P, c)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
