% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.calc.R
\name{res.calc}
\alias{res.calc}
\title{Residual calculation}
\usage{
res.calc(x, Y, a)
}
\arguments{
\item{Y}{nx1 response vector}

\item{a}{nxq confounding matrix}

\item{X}{nxp predictor matrix}
}
\description{
Residual calculation
}
\examples{
res.calc(X,Y,a)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
