% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respkernel.R
\name{respkernel}
\alias{respkernel}
\title{Generates response kernel.  Can be either linear or delta kernel}
\usage{
respkernel(Y, n, kernel)
}
\arguments{
\item{Y}{Response vector}

\item{n}{number of observations in response vector Y}

\item{kernel}{Type of kernel.  Either linear or delta}
}
\description{
Generates response kernel.  Can be either linear or delta kernel
}
\examples{
respkernel(Y, n, kernel)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
