% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfast.mat.eig.sym.decomp.R
\name{rfast.mat.eig.sym.decomp}
\alias{rfast.mat.eig.sym.decomp}
\title{Eigendecomposition using residualization process with Rfast coding}
\usage{
rfast.mat.eig.sym.decomp(X.star, X.tilde, A, L, H, lambda, npc)
}
\arguments{
\item{X.star}{Data matrix residualized with respect to confounding variables}

\item{X.tilde}{Data matrix residualized with respect to response variable}

\item{A}{kernel matrix of confounders}

\item{L}{kernel matrix of response}

\item{H}{centering matrix}

\item{lambda}{penalty term for confounding variable adjustment}

\item{npc}{number of principl components}
}
\description{
Eigendecomposition using residualization process with Rfast coding
}
\examples{
rfast.mat.eig.sym.decomp(X.star,X.tilde,A,L,H,lambda,npc)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
