% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpls.deflation.regular.R
\name{simpls.deflation.regular}
\alias{simpls.deflation.regular}
\title{SIMPLS deflation step}
\usage{
simpls.deflation.regular(X, V, n, p)
}
\arguments{
\item{X}{nxp data matrix}

\item{V}{matrix of principal component vectors}

\item{n}{number of observations}

\item{p}{number of predictors}
}
\description{
SIMPLS deflation step
}
\examples{
simpls.deflation.regular(X,V,n,p)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
