% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.vector.decomp.R
\name{single.vector.decomp}
\alias{single.vector.decomp}
\title{Extract single vector for SIMPLS algorithm}
\usage{
single.vector.decomp(X, H, Ky, Ka, lambda)
}
\arguments{
\item{X}{nxp predictor matrix}

\item{H}{centering matrix}

\item{Ky}{response kernel matrix}

\item{Ka}{confounding kernel matrix}
}
\description{
Extract single vector for SIMPLS algorithm
}
\examples{
single.vector.decomp <- function(X,H,Ky,Ka,lambda)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
