% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sspca.orthog.R
\name{sspca.orthog}
\alias{sspca.orthog}
\title{Orthogonalize design matrix to extracted principal component vector}
\usage{
sspca.orthog(X, v, nobs, p)
}
\arguments{
\item{X}{nxp predictor matrix}

\item{v}{principal component vectors}

\item{nobs}{number of responses}

\item{p}{number of predictors}
}
\description{
Orthogonalize design matrix to extracted principal component vector
}
\examples{
sspca.orthog(X,v,nobs,p)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Supervised}
