% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topthresfilter.R
\name{topthresfilter}
\alias{topthresfilter}
\title{This function converts very low magnitude loadings into zero}
\usage{
topthresfilter(v, top, alpha)
}
\arguments{
\item{v}{vector to be converted}

\item{top}{theshold value}

\item{alpha}{theshold limit}
}
\description{
This function converts very low magnitude loadings into zero
}
\examples{
topthresfilter(v, top, alpha)
}
\keyword{Adjusted}
\keyword{Analysis}
\keyword{Component}
\keyword{Confounding}
\keyword{Principal}
\keyword{Sparse}
\keyword{Supervised}
